Finish setting up data collection launch
