+++ phoenix_gazebo/launch/inference.launch.py
@@ -1,0 +1,1 @@
+#TODO create when inference is done
